update pack script to use path_utils
